refactor: move runtime functions out of APIs
