+++ apis/apiextensions/v1/composition_transforms_test.go
@@ -338,107 +338,6 @@ func TestTransformValidate(t *testing.T) {
 	}
 }
 
-func TestConvertTransformGetConversionFunc(t *testing.T) {
-	type args struct {
-		ct   *ConvertTransform
-		from TransformIOType
-	}
-	type want struct {
-		err error
-	}
-	cases := map[string]struct {
-		reason string
-		args   args
-		want   want
-	}{
-		"IntToString": {
-			reason: "Int to String should be valid",
-			args: args{
-				ct: &ConvertTransform{
-					ToType: TransformIOTypeString,
-				},
-				from: TransformIOTypeInt,
-			},
-		},
-		"IntToInt": {
-			reason: "Int to Int should be valid",
-			args: args{
-				ct: &ConvertTransform{
-					ToType: TransformIOTypeInt,
-				},
-				from: TransformIOTypeInt,
-			},
-		},
-		"IntToInt64": {
-			reason: "Int to Int64 should be valid",
-			args: args{
-				ct: &ConvertTransform{
-					ToType: TransformIOTypeInt,
-				},
-				from: TransformIOTypeInt64,
-			},
-		},
-		"Int64ToInt": {
-			reason: "Int64 to Int should be valid",
-			args: args{
-				ct: &ConvertTransform{
-					ToType: TransformIOTypeInt64,
-				},
-				from: TransformIOTypeInt,
-			},
-		},
-		"IntToFloat": {
-			reason: "Int to Float should be valid",
-			args: args{
-				ct: &ConvertTransform{
-					ToType: TransformIOTypeInt,
-				},
-				from: TransformIOTypeFloat64,
-			},
-		},
-		"IntToBool": {
-			reason: "Int to Bool should be valid",
-			args: args{
-				ct: &ConvertTransform{
-					ToType: TransformIOTypeInt,
-				},
-				from: TransformIOTypeBool,
-			},
-		},
-		"StringToIntInvalidFormat": {
-			reason: "String to Int with invalid format should be invalid",
-			args: args{
-				ct: &ConvertTransform{
-					ToType: TransformIOTypeInt,
-					Format: &[]ConvertTransformFormat{"wrong"}[0],
-				},
-				from: TransformIOTypeString,
-			},
-			want: want{
-				err: fmt.Errorf("conversion from string to int64 is not supported with format wrong"),
-			},
-		},
-		"IntToIntInvalidFormat": {
-			reason: "Int to Int, invalid format ignored because it is the same type",
-			args: args{
-				ct: &ConvertTransform{
-					ToType: TransformIOTypeInt,
-					Format: &[]ConvertTransformFormat{"wrong"}[0],
-				},
-				from: TransformIOTypeInt,
-			},
-		},
-	}
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			_, err := tc.args.ct.GetConversionFunc(tc.args.from)
-			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
-				t.Errorf("%s\nGetConversionFunc(...): -want, +got:\n%s", tc.reason, diff)
-			}
-		})
-	}
-}
-
 func TestTransformGetOutputType(t *testing.T) {
 	type args struct {
 		transform *Transform
